+++ resources/release-notes.md
@@ -6,4 +6,4 @@
 - Remove `mag.Science/computePSD` method (use `mag.psd` instead)
 - Fix issues with plotting IMAP HK and timestamp analysis with one sensor data missing
 - Fix issues with methods of `mag.Science` with no data
-- Add tests for `mag.Spectrum`
+- Add tests for `mag.Spectrum` and `mag.spectrogram`